Move reository selection code to own module
